backend/requirements.txt
# Backend TDR Generator - Dependencias
# Instalar con: pip install -r requirements.txt

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17

# Base de datos vectorial
chromadb>=0.5.0

# Embeddings y LLM
google-genai>=1.0.0
sentence-transformers>=3.0.0

# Utilidades
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
httpx>=0.28.0

# CORS y JWT
python-jose>=3.3.0
PyJWT>=2.8.0

# PostgreSQL
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

